Split the file into two, so we now have a file with the Blocks and one with the Configuration of the datatype
